Allow binding outgoing address
There is cases where it may be desireable to bind the source address of
outgoing connections.
This diff adds support for it.

When using UDP, this will cause the outgoing connection to listen on
IP:port vs *:port.
When using TCP, the outgoing connection is going to use the provide IP
vs letting the OS do it.

Tested by running a local unbound instance. When not binding to eth0 IP,
connections would be coming from 127.0.0.1, when using
--outgoing-address, connections would be coming from the provided IP.

Likewise with ipv6, tested using --outgoing-address=[::1] against a
resolver listening on ::1 and --outgoing-address=[::1]:portnumber.
This confirmed we would indeed bind to the provided port. Binding to 1
specific port for outgoing connection is not desirable, but provided
users only provide the IP to bind to, we can piggyback on the existing
`sockaddr_from_ip_and_port` function.