Use sodium_malloc() for the DNS query/response buffers
They are currently allocated on the stack. Allocate them using
sodium_malloc() instead.

I'm not confident in the way the response buffer is used. What is
written into it relative to its size is not checked everywhere
(for example in the function that serves certificates).

Having them on the heap makes exploitation more difficult if
something goes wrong. sodium_malloc() also adds a guard page,
which can be useful. I'd rather have the proxy crash due to
hitting the guard page than hit the stack.